requirements.txt
fastapi[all]==0.75.2
tortoise-orm[aiomysql]==0.19.2
aerich==0.6.3
typer[all]==0.6.1
passlib==1.7.4
python-jose[cryptography]==3.3.0
aioredis[hiredis]==2.0.1
simpel_captcha==0.1.3
gunicorn==20.1.0
uvicorn==0.17.6
pytest==7.1.2
httpx==0.23.0